Handle unknown queues
Log a warning, but don't error, when we aren't able to match the task to
a queue. This may happen, for instance, if there is a race condition,
where no task for a queue is running when `active_queues` is called, and
therefore it isn't present in that call, but when the active tasks are
queried, there is a task on that queue.